Separate functions that don't belong to CastSpell class
