+++ apps/openmw/mwmechanics/autocalcspell.cpp
@@ -1,5 +1,4 @@
 #include "autocalcspell.hpp"
-#include "spellcasting.hpp"
 
 #include <limits>
 
@@ -8,6 +7,7 @@
 #include "../mwbase/world.hpp"
 #include "../mwbase/environment.hpp"
 
+#include "spellutil.hpp"
 
 namespace MWMechanics
 {
