+++ apps/openmw/mwmechanics/spellcasting.cpp
@@ -1,8 +1,5 @@
 #include "spellcasting.hpp"
 
-#include <limits>
-#include <iomanip>
-
 #include <components/misc/constants.hpp>
 #include <components/misc/rng.hpp>
 #include <components/settings/settings.hpp>
@@ -29,185 +26,11 @@
 #include "weapontype.hpp"
 #include "summoning.hpp"
 #include "spellresistance.hpp"
+#include "spellutil.hpp"
+#include "tickableeffects.hpp"
 
 namespace MWMechanics
 {
-    ESM::Skill::SkillEnum spellSchoolToSkill(int school)
-    {
-        static const std::array<ESM::Skill::SkillEnum, 6> schoolSkillArray
-        {
-            ESM::Skill::Alteration, ESM::Skill::Conjuration, ESM::Skill::Destruction,
-            ESM::Skill::Illusion, ESM::Skill::Mysticism, ESM::Skill::Restoration
-        };
-        return schoolSkillArray.at(school);
-    }
-
-    float calcEffectCost(const ESM::ENAMstruct& effect, const ESM::MagicEffect* magicEffect)
-    {
-        if (!magicEffect)
-            magicEffect = MWBase::Environment::get().getWorld()->getStore().get<ESM::MagicEffect>().find(effect.mEffectID);
-        bool hasMagnitude = !(magicEffect->mData.mFlags & ESM::MagicEffect::NoMagnitude);
-        bool hasDuration = !(magicEffect->mData.mFlags & ESM::MagicEffect::NoDuration);
-        int minMagn = hasMagnitude ? effect.mMagnMin : 1;
-        int maxMagn = hasMagnitude ? effect.mMagnMax : 1;
-        int duration = hasDuration ? effect.mDuration : 1;
-        static const float fEffectCostMult = MWBase::Environment::get().getWorld()->getStore()
-            .get<ESM::GameSetting>().find("fEffectCostMult")->mValue.getFloat();
-
-        float x = 0.5 * (std::max(1, minMagn) + std::max(1, maxMagn));
-        x *= 0.1 * magicEffect->mData.mBaseCost;
-        x *= 1 + duration;
-        x += 0.05 * std::max(1, effect.mArea) * magicEffect->mData.mBaseCost;
-
-        return x * fEffectCostMult;
-    }
-
-    float calcSpellBaseSuccessChance (const ESM::Spell* spell, const MWWorld::Ptr& actor, int* effectiveSchool)
-    {
-        // Morrowind for some reason uses a formula slightly different from magicka cost calculation
-        float y = std::numeric_limits<float>::max();
-        float lowestSkill = 0;
-
-        for (const ESM::ENAMstruct& effect : spell->mEffects.mList)
-        {
-            float x = static_cast<float>(effect.mDuration);
-            const auto magicEffect = MWBase::Environment::get().getWorld()->getStore().get<ESM::MagicEffect>().find(effect.mEffectID);
-
-            if (!(magicEffect->mData.mFlags & ESM::MagicEffect::AppliedOnce))
-                x = std::max(1.f, x);
-
-            x *= 0.1f * magicEffect->mData.mBaseCost;
-            x *= 0.5f * (effect.mMagnMin + effect.mMagnMax);
-            x += effect.mArea * 0.05f * magicEffect->mData.mBaseCost;
-            if (effect.mRange == ESM::RT_Target)
-                x *= 1.5f;
-            static const float fEffectCostMult = MWBase::Environment::get().getWorld()->getStore().get<ESM::GameSetting>().find(
-                        "fEffectCostMult")->mValue.getFloat();
-            x *= fEffectCostMult;
-
-            float s = 2.0f * actor.getClass().getSkill(actor, spellSchoolToSkill(magicEffect->mData.mSchool));
-            if (s - x < y)
-            {
-                y = s - x;
-                if (effectiveSchool)
-                    *effectiveSchool = magicEffect->mData.mSchool;
-                lowestSkill = s;
-            }
-        }
-
-        CreatureStats& stats = actor.getClass().getCreatureStats(actor);
-
-        int actorWillpower = stats.getAttribute(ESM::Attribute::Willpower).getModified();
-        int actorLuck = stats.getAttribute(ESM::Attribute::Luck).getModified();
-
-        float castChance = (lowestSkill - spell->mData.mCost + 0.2f * actorWillpower + 0.1f * actorLuck);
-
-        return castChance;
-    }
-
-    float getSpellSuccessChance (const ESM::Spell* spell, const MWWorld::Ptr& actor, int* effectiveSchool, bool cap, bool checkMagicka)
-    {
-        bool godmode = actor == MWMechanics::getPlayer() && MWBase::Environment::get().getWorld()->getGodModeState();
-
-        CreatureStats& stats = actor.getClass().getCreatureStats(actor);
-
-        float castBonus = -stats.getMagicEffects().get(ESM::MagicEffect::Sound).getMagnitude();
-
-        float castChance = calcSpellBaseSuccessChance(spell, actor, effectiveSchool) + castBonus;
-        castChance *= stats.getFatigueTerm();
-
-        if (stats.getMagicEffects().get(ESM::MagicEffect::Silence).getMagnitude()&& !godmode)
-            return 0;
-
-        if (spell->mData.mType == ESM::Spell::ST_Power)
-            return stats.getSpells().canUsePower(spell) ? 100 : 0;
-
-        if (spell->mData.mType != ESM::Spell::ST_Spell)
-            return 100;
-
-        if (checkMagicka && stats.getMagicka().getCurrent() < spell->mData.mCost && !godmode)
-            return 0;
-
-        if (spell->mData.mFlags & ESM::Spell::F_Always)
-            return 100;
-
-        if (godmode)
-            return 100;
-
-        return std::max(0.f, cap ? std::min(100.f, castChance) : castChance);
-    }
-
-    float getSpellSuccessChance (const std::string& spellId, const MWWorld::Ptr& actor, int* effectiveSchool, bool cap, bool checkMagicka)
-    {
-        if (const auto spell = MWBase::Environment::get().getWorld()->getStore().get<ESM::Spell>().search(spellId))
-            return getSpellSuccessChance(spell, actor, effectiveSchool, cap, checkMagicka);
-        return 0.f;
-    }
-
-    int getSpellSchool(const std::string& spellId, const MWWorld::Ptr& actor)
-    {
-        int school = 0;
-        getSpellSuccessChance(spellId, actor, &school);
-        return school;
-    }
-
-    int getSpellSchool(const ESM::Spell* spell, const MWWorld::Ptr& actor)
-    {
-        int school = 0;
-        getSpellSuccessChance(spell, actor, &school);
-        return school;
-    }
-
-    bool spellIncreasesSkill(const ESM::Spell *spell)
-    {
-        return spell->mData.mType == ESM::Spell::ST_Spell && !(spell->mData.mFlags & ESM::Spell::F_Always);
-    }
-
-    bool spellIncreasesSkill(const std::string &spellId)
-    {
-        const auto spell = MWBase::Environment::get().getWorld()->getStore().get<ESM::Spell>().search(spellId);
-        return spell && spellIncreasesSkill(spell);
-    }
-
-    /// Check if the given effect can be applied to the target. If \a castByPlayer, emits a message box on failure.
-    bool checkEffectTarget (int effectId, const MWWorld::Ptr& target, const MWWorld::Ptr& caster, bool castByPlayer)
-    {
-        switch (effectId)
-        {
-            case ESM::MagicEffect::Levitate:
-                if (!MWBase::Environment::get().getWorld()->isLevitationEnabled())
-                {
-                    if (castByPlayer)
-                        MWBase::Environment::get().getWindowManager()->messageBox("#{sLevitateDisabled}");
-                    return false;
-                }
-                break;
-            case ESM::MagicEffect::Soultrap:
-                if (!target.getClass().isNpc() // no messagebox for NPCs
-                     && (target.getTypeName() == typeid(ESM::Creature).name() && target.get<ESM::Creature>()->mBase->mData.mSoul == 0))
-                {
-                    if (castByPlayer)
-                        MWBase::Environment::get().getWindowManager()->messageBox("#{sMagicInvalidTarget}");
-                    return true; // must still apply to get visual effect and have target regard it as attack
-                }
-                break;
-            case ESM::MagicEffect::WaterWalking:
-                if (target.getClass().isPureWaterCreature(target) && MWBase::Environment::get().getWorld()->isSwimming(target))
-                    return false;
-
-                MWBase::World *world = MWBase::Environment::get().getWorld();
-
-                if (!world->isWaterWalkingCastableOnTarget(target))
-                {
-                    if (castByPlayer && caster == target)
-                        MWBase::Environment::get().getWindowManager()->messageBox ("#{sMagicInvalidEffect}");
-                    return false;
-                }
-                break;
-        }
-        return true;
-    }
-
     class GetAbsorptionProbability : public MWMechanics::EffectSourceVisitor
     {
     public:
@@ -1015,233 +838,5 @@ namespace MWMechanics
     bool CastSpell::spellIncreasesSkill()
     {
         return !mManualSpell && MWMechanics::spellIncreasesSkill(mId);
-    }
-
-    int getEffectiveEnchantmentCastCost(float castCost, const MWWorld::Ptr &actor)
-    {
-        /*
-         * Each point of enchant skill above/under 10 subtracts/adds
-         * one percent of enchantment cost while minimum is 1.
-         */
-        int eSkill = actor.getClass().getSkill(actor, ESM::Skill::Enchant);
-        const float result = castCost - (castCost / 100) * (eSkill - 10);
-
-        return static_cast<int>((result < 1) ? 1 : result);
-    }
-
-    bool disintegrateSlot (MWWorld::Ptr ptr, int slot, float disintegrate)
-    {
-        if (ptr.getClass().hasInventoryStore(ptr))
-        {
-            MWWorld::InventoryStore& inv = ptr.getClass().getInventoryStore(ptr);
-            MWWorld::ContainerStoreIterator item = inv.getSlot(slot);
-
-            if (item != inv.end() && (item.getType() == MWWorld::ContainerStore::Type_Armor || item.getType() == MWWorld::ContainerStore::Type_Weapon))
-            {
-                if (!item->getClass().hasItemHealth(*item))
-                    return false;
-                int charge = item->getClass().getItemHealth(*item);
-
-                if (charge == 0)
-                    return false;
-
-                // Store remainder of disintegrate amount (automatically subtracted if > 1)
-                item->getCellRef().applyChargeRemainderToBeSubtracted(disintegrate - std::floor(disintegrate));
-
-                charge = item->getClass().getItemHealth(*item);
-                charge -= std::min(static_cast<int>(disintegrate), charge);
-                item->getCellRef().setCharge(charge);
-
-                if (charge == 0)
-                {
-                    // Will unequip the broken item and try to find a replacement
-                    if (ptr != getPlayer())
-                        inv.autoEquip(ptr);
-                    else
-                        inv.unequipItem(*item, ptr);
-                }
-
-                return true;
-            }
-        }
-        return false;
-    }
-
-    void adjustDynamicStat(CreatureStats& creatureStats, int index, float magnitude, bool allowDecreaseBelowZero = false)
-    {
-        DynamicStat<float> stat = creatureStats.getDynamic(index);
-        stat.setCurrent(stat.getCurrent() + magnitude, allowDecreaseBelowZero);
-        creatureStats.setDynamic(index, stat);
-    }
-
-    bool effectTick(CreatureStats& creatureStats, const MWWorld::Ptr& actor, const EffectKey &effectKey, float magnitude)
-    {
-        if (magnitude == 0.f)
-            return false;
-
-        bool receivedMagicDamage = false;
-
-        switch (effectKey.mId)
-        {
-        case ESM::MagicEffect::DamageAttribute:
-        {
-            AttributeValue attr = creatureStats.getAttribute(effectKey.mArg);
-            attr.damage(magnitude);
-            creatureStats.setAttribute(effectKey.mArg, attr);
-            break;
-        }
-        case ESM::MagicEffect::RestoreAttribute:
-        {
-            AttributeValue attr = creatureStats.getAttribute(effectKey.mArg);
-            attr.restore(magnitude);
-            creatureStats.setAttribute(effectKey.mArg, attr);
-            break;
-        }
-        case ESM::MagicEffect::RestoreHealth:
-        case ESM::MagicEffect::RestoreMagicka:
-        case ESM::MagicEffect::RestoreFatigue:
-            adjustDynamicStat(creatureStats, effectKey.mId-ESM::MagicEffect::RestoreHealth, magnitude);
-            break;
-        case ESM::MagicEffect::DamageHealth:
-            receivedMagicDamage = true;
-            adjustDynamicStat(creatureStats, effectKey.mId-ESM::MagicEffect::DamageHealth, -magnitude);
-            break;
-
-        case ESM::MagicEffect::DamageMagicka:
-        case ESM::MagicEffect::DamageFatigue:
-        {
-            int index = effectKey.mId-ESM::MagicEffect::DamageHealth;
-            static const bool uncappedDamageFatigue = Settings::Manager::getBool("uncapped damage fatigue", "Game");
-            adjustDynamicStat(creatureStats, index, -magnitude, index == 2 && uncappedDamageFatigue);
-            break;
-        }
-        case ESM::MagicEffect::AbsorbHealth:
-            if (magnitude > 0.f)
-                receivedMagicDamage = true;
-            adjustDynamicStat(creatureStats, effectKey.mId-ESM::MagicEffect::AbsorbHealth, -magnitude);
-
-            break;
-
-        case ESM::MagicEffect::AbsorbMagicka:
-        case ESM::MagicEffect::AbsorbFatigue:
-            adjustDynamicStat(creatureStats, effectKey.mId-ESM::MagicEffect::AbsorbHealth, -magnitude);
-            break;
-
-        case ESM::MagicEffect::DisintegrateArmor:
-        {
-            // According to UESP
-            int priorities[] = {
-                MWWorld::InventoryStore::Slot_CarriedLeft,
-                MWWorld::InventoryStore::Slot_Cuirass,
-                MWWorld::InventoryStore::Slot_LeftPauldron,
-                MWWorld::InventoryStore::Slot_RightPauldron,
-                MWWorld::InventoryStore::Slot_LeftGauntlet,
-                MWWorld::InventoryStore::Slot_RightGauntlet,
-                MWWorld::InventoryStore::Slot_Helmet,
-                MWWorld::InventoryStore::Slot_Greaves,
-                MWWorld::InventoryStore::Slot_Boots
-            };
-
-            for (unsigned int i=0; i<sizeof(priorities)/sizeof(int); ++i)
-            {
-                if (disintegrateSlot(actor, priorities[i], magnitude))
-                    break;
-            }
-
-            break;
-        }
-        case ESM::MagicEffect::DisintegrateWeapon:
-            disintegrateSlot(actor, MWWorld::InventoryStore::Slot_CarriedRight, magnitude);
-            break;
-
-        case ESM::MagicEffect::SunDamage:
-        {
-            // isInCell shouldn't be needed, but updateActor called during game start
-            if (!actor.isInCell() || !actor.getCell()->isExterior())
-                break;
-            float time = MWBase::Environment::get().getWorld()->getTimeStamp().getHour();
-            float timeDiff = std::min(7.f, std::max(0.f, std::abs(time - 13)));
-            float damageScale = 1.f - timeDiff / 7.f;
-            // When cloudy, the sun damage effect is halved
-            static float fMagicSunBlockedMult = MWBase::Environment::get().getWorld()->getStore().get<ESM::GameSetting>().find(
-                        "fMagicSunBlockedMult")->mValue.getFloat();
-
-            int weather = MWBase::Environment::get().getWorld()->getCurrentWeather();
-            if (weather > 1)
-                damageScale *= fMagicSunBlockedMult;
-
-            adjustDynamicStat(creatureStats, 0, -magnitude * damageScale);
-            if (magnitude * damageScale > 0.f)
-                receivedMagicDamage = true;
-
-            break;
-        }
-
-        case ESM::MagicEffect::FireDamage:
-        case ESM::MagicEffect::ShockDamage:
-        case ESM::MagicEffect::FrostDamage:
-        case ESM::MagicEffect::Poison:
-        {
-            adjustDynamicStat(creatureStats, 0, -magnitude);
-            receivedMagicDamage = true;
-            break;
-        }
-
-        case ESM::MagicEffect::DamageSkill:
-        case ESM::MagicEffect::RestoreSkill:
-        {
-            if (!actor.getClass().isNpc())
-                break;
-            NpcStats &npcStats = actor.getClass().getNpcStats(actor);
-            SkillValue& skill = npcStats.getSkill(effectKey.mArg);
-            if (effectKey.mId == ESM::MagicEffect::RestoreSkill)
-                skill.restore(magnitude);
-            else
-                skill.damage(magnitude);
-            break;
-        }
-
-        case ESM::MagicEffect::CurePoison:
-            actor.getClass().getCreatureStats(actor).getActiveSpells().purgeEffect(ESM::MagicEffect::Poison);
-            break;
-        case ESM::MagicEffect::CureParalyzation:
-            actor.getClass().getCreatureStats(actor).getActiveSpells().purgeEffect(ESM::MagicEffect::Paralyze);
-            break;
-        case ESM::MagicEffect::CureCommonDisease:
-            actor.getClass().getCreatureStats(actor).getSpells().purgeCommonDisease();
-            break;
-        case ESM::MagicEffect::CureBlightDisease:
-            actor.getClass().getCreatureStats(actor).getSpells().purgeBlightDisease();
-            break;
-        case ESM::MagicEffect::CureCorprusDisease:
-            actor.getClass().getCreatureStats(actor).getSpells().purgeCorprusDisease();
-            break;
-        case ESM::MagicEffect::RemoveCurse:
-            actor.getClass().getCreatureStats(actor).getSpells().purgeCurses();
-            break;
-        default:
-            return false;
-        }
-
-        if (receivedMagicDamage && actor == getPlayer())
-            MWBase::Environment::get().getWindowManager()->activateHitOverlay(false);
-        return true;
-    }
-
-    void ApplyLoopingParticlesVisitor::visit (MWMechanics::EffectKey key,
-                        const std::string& /*sourceName*/, const std::string& /*sourceId*/, int /*casterActorId*/,
-                        float /*magnitude*/, float /*remainingTime*/, float /*totalTime*/)
-    {
-        const auto magicEffect = MWBase::Environment::get().getWorld()->getStore().get<ESM::MagicEffect>().find(key.mId);
-        if ((magicEffect->mData.mFlags & ESM::MagicEffect::ContinuousVfx) == 0)
-            return;
-        const ESM::Static* castStatic;
-        if (!magicEffect->mHit.empty())
-            castStatic = MWBase::Environment::get().getWorld()->getStore().get<ESM::Static>().find (magicEffect->mHit);
-        else
-            castStatic = MWBase::Environment::get().getWorld()->getStore().get<ESM::Static>().find ("VFX_DefaultHit");
-        MWRender::Animation* anim = MWBase::Environment::get().getWorld()->getAnimation(mActor);
-        if (anim && !castStatic->mModel.empty())
-            anim->addEffect("meshes\\" + castStatic->mModel, magicEffect->mIndex, /*loop*/true, "", magicEffect->mParticle);
     }
 }
